Refactor stereochemistry guidance configuration
