+++ openfold3/projects/of3_all_atom/config/dataset_configs.py
@@ -47,7 +47,7 @@ from openfold3.core.config.config_utils import (
     FilePathOrNone,
     deep_update,
 )
-from openfold3.core.config.ligand_stereochemistry import (
+from openfold3.core.config.ligand_stereochemistry_config import (
     LigandStereochemistryGuidanceSettings,
 )
 from openfold3.core.config.pocket_sampling_config import PocketSamplingSettings
@@ -370,12 +370,15 @@ class TrainingDatasetSpec(DatasetSpec):
 
 
 class InferenceDatasetConfigKwargs(BaseModel):
-    """Class to hold msa, template, and pocket sampling kwargs for inference pipeline"""
+    """Class to hold inference dataset configuration."""
 
     ccd_file_path: FilePathOrNone = None
     msa: MSASettings = MSASettings(subsample_main=False)
     template: TemplateSettings = TemplateSettings(take_top_k=True)
     pocket_sampling: PocketSamplingSettings = PocketSamplingSettings()
+    ligand_stereochemistry_guidance: LigandStereochemistryGuidanceSettings = (
+        LigandStereochemistryGuidanceSettings()
+    )
 
 
 class InferenceJobConfig(BaseModel):
@@ -388,7 +391,7 @@ class InferenceJobConfig(BaseModel):
     template: TemplateSettings = TemplateSettings()
     template_preprocessor_settings: TemplatePreprocessorSettings
     pocket_sampling: PocketSamplingSettings = PocketSamplingSettings()
-    ligand_stereochemistry_guidance_settings: LigandStereochemistryGuidanceSettings = (
+    ligand_stereochemistry_guidance: LigandStereochemistryGuidanceSettings = (
         LigandStereochemistryGuidanceSettings()
     )
